feat: home routes
