+++ src/routes/Home.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Navigation from '../components/Navigation';
+import HomeMyInfo from '../components/HomeMyInfo';
+const Home = () => {
+  return (
+    <section className="navigatior">
+      <Navigation />
+      <HomeMyInfo />
+    </section>
+  );
+};
+
+export default Home;
